# docs/openapi.yaml
openapi: 3.1.0
info:
  title: IdentityHub REST API
  version: "1.0.0"
  description: |
    REST API for IdentityHub — turn Non-Human Identity (NHI) findings into
    tracked **Jira** tickets from scanners, CI/CD, or the web UI.

    ## Authentication

    Two mechanisms are supported:

    - **Session cookie** (`session`) — issued by `POST /api/auth/login` (or
      `/register`). Used by the web UI and required by every endpoint except
      where noted.
    - **API key** (`X-API-Key` header) — a key with the `nhk_` prefix, generated
      from `POST /api/keys`. Accepted **only** on the ticket endpoints
      (`POST /api/tickets`, `GET /api/tickets`), which is where automation runs.

    Endpoints that accept either list both schemes; a request needs to satisfy
    just one of them.

    ## Errors

    Errors return `{ "error": "<name>", "message": "<human message>" }` with the
    matching HTTP status. Jira upstream problems surface as `502` (Jira error) or
    `503` (Jira unreachable).
servers:
  - url: /
    description: Same origin as the running app (e.g. http://localhost:5000)
...
paths:
...
  /api/tickets:
    post:
      tags: [Tickets]
      summary: Create a Jira ticket
      description: |
        Files an NHI finding as a Jira issue. Requires a connected Jira
        workspace. Tickets are **not** persisted locally — Jira is the source of
        truth.
      security:
        - sessionCookie: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketCreateRequest"
      responses:
        "201":
          description: Ticket created in Jira.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketCreated"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/JiraError"
        "503":
          $ref: "#/components/responses/JiraUnavailable"
    get:
      tags: [Tickets]
      summary: List recent tickets
      description: |
        Recent tickets read **live from Jira** every request. If no Jira
        workspace is connected, returns an empty list with `unavailable: false`.
        If Jira is unreachable, returns an empty list with `unavailable: true`.
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - name: project_key
          in: query
          description: Restrict to a single project (uppercased server-side).
          required: false
          schema:
            type: string
            example: NHI
        - name: limit
          in: query
          description: Max tickets to return (clamped to 1–100).
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Recent tickets.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Ticket"
                  unavailable:
                    type: boolean
                    description: True when Jira could not be reached.
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie set by login/register.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: An API key with the `nhk_` prefix. Ticket endpoints only.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error name (e.g. ValidationError, NotFoundError).
        message:
          type: string
          description: Human-readable message.
      required: [error, message]
...
    TicketCreateRequest:
      type: object
      properties:
        project_key:
          type: string
          description: Jira project key (uppercased server-side).
          example: NHI
        summary:
          type: string
          maxLength: 500
        description:
          type: string
      required: [project_key, summary]
    TicketCreated:
      type: object
      properties:
        jira_key:
          type: string
          example: NHI-42
        jira_id:
          type: string
        project_key:
          type: string
        summary:
          type: string
        description:
          type: [string, "null"]
        source:
          type: string
          enum: [web, api]
          description: web = session request; api = X-API-Key request.
      required: [jira_key, jira_id, project_key, summary, source]
    Ticket:
      type: object
      properties:
        id:
          type: [integer, "null"]
        jira_key:
          type: string
          example: NHI-42
        jira_id:
          type: string
        project_key:
          type: string
        summary:
          type: string
        description:
          type: [string, "null"]
        source:
          type: string
          example: jira
        created_at:
          type: [string, "null"]
          format: date-time
...
  responses:
...
    Unauthorized:
      description: Authentication required or session expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    JiraError:
      description: Jira returned an error (bad credentials or invalid response).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    JiraUnavailable:
      description: Jira could not be reached (connection failure, timeout, or 5xx).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"